1.Add MapUtil、assets and pipe 2.Add addPipeTrack in MapModel

+++ app/src/main/java/com/barisetech/www/workmanage/bean/map/MapPosition.java
@@ -7,14 +7,14 @@ public class MapPosition {
     /**
      * 纬度
      */
-    public float latitude;
+    public double latitude;
 
     /**
      * 经度
      */
-    public float longitude;
+    public double longitude;
 
-    public MapPosition(float latitude, float longitude) {
+    public MapPosition(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
